requirements.txt
xgboost~=1.4.1
numpy~=1.21
pandas~=1.2.0
scikit-learn~=0.24.1
scipy~=1.6.1
docopt~=0.6.2
matplotlib~=3.3.3
catboost~=0.24.4
lightgbm~=3.1.1
requests~=2.25.1
js2py~=0.70
sklearn~=0.0
tqdm~=4.59.0
pyenchant~=3.2.0
beautifulsoup4~=4.9.3
Levenshtein~=0.15.0